llvm20 linting fixes, containerize linux builds

+++ src/unrealsdk/game/selector.cpp
@@ -37,7 +37,7 @@ using all_known_games = std::tuple<BL2Hook, TPSHook>;
  * @tparam i Index of the game class being tested this iteration. Picked up automatically.
  * @param executable The executable name to match against.
  */
-template <int i = 0>
+template <size_t i = 0>
 std::unique_ptr<AbstractHook> find_correct_hook(std::string_view executable) {
     if constexpr (i >= std::tuple_size_v<all_known_games>) {
         throw std::runtime_error("Failed to find compatible game hook!");
